+++ js/cyclone.js
@@ -1,10 +1,26 @@
 /**
  * Cyclone app
  *
- * Created by Cyclodex
+ * Created by Fabian Gander | Cyclodex
  */
 
-// Here we require all the code needed so webpack knows what to load.
+// Load angular libraries
+require('angular');
+require('angular-route');
+require('angular-animate');
+require('angular-aria');
+require('angular-messages');
+require('angular-material');
+require('angular-clipboard');
+require('angular-moment');
+
+// Firebase
+var firebase = require('firebase/app');
+require('firebase/auth');
+require('firebase/database');
+require('angularfire');
+
+// Cyclone app files
 require('./config/config.js');
 require('./cyclone.module.js');
 require('./components/auth.factory.js');